pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imessage-max"
version = "0.1.0"
description = "Intent-aligned MCP server for iMessage"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Rob Dezendorf" }
]
dependencies = [
    "fastmcp>=0.4,<3",
    "pyobjc-framework-Contacts>=12.1",
    "phonenumbers>=8.13",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
imessage-max = "imessage_max.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/imessage_max"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
